split out indicators to its own module
